cssom-view/elementFromPoint.html

<!DOCTYPE html>
<title>cssom-view - elementFromPoint</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  .size {
    width:100px;
    height:100px;
  }
  .overlay {
    position:absolute;
    top:109px;
    pointer-events:none;
  }
  .purple {
    background-color: rebeccapurple;
  }
  .red {
    background-color: red;
  }
  .green {
    background-color: green;
  }
</style>
<body>
  <div id='purple' class="size purple" >&nbsp;</div>
  <div id='red' class="size red">&nbsp;</div>
  <div id='green' class="size overlay green">&nbsp;</div>
  <iframe id=iframe-1 src="iframe.html" style='display:none;position:absolute; left:300px;'></iframe>
  <iframe id=iframe-2 src="iframe.html" width="" height=""></iframe>
  <svg id=squiggle xmlns="http://www.w3.org/2000/svg" height="98" width="581" viewBox="0 0 581 98">
    <path stroke-dashoffset="0.00" stroke-dasharray="" d="M62.9 14.9c-25-7.74-56.6 4.8-60.4 24.3-3.73 19.6 21.6 35 39.6 37.6 42.8 6.2 72.9-53.4 116-58.9 65-18.2 191 101 215 28.8 5-16.7-7-49.1-34-44-34 11.5-31 46.5-14 69.3 9.38 12.6 24.2 20.6 39.8 22.9 91.4 9.05 102-98.9 176-86.7 18.8 3.81 33 17.3 36.7 34.6 2.01 10.2.124 21.1-5.18 30.1" stroke="#000" stroke-width="4.3" fill="none">
    </path>
  </svg>
  <script>
    "use strict";
    test(function () {
      debugger;
                  assert_equals(null, document.elementFromPoint(-1, -1),
                    "both co-ordinates passed in are negative so should have returned a null");
                  assert_equals(null, document.elementFromPoint(-1, 1),
                    "x co-ordinates passed in are negative so should have returned a null");
                  assert_equals(null, document.elementFromPoint(1, -1),
                    "y co-ordinates passed in are negative so should have returned a null");
                }, "Negative co-ordinates");

    test(function () {
                  var viewportX = window.innerWidth;
                  var viewportY = window.innerHeight;
                  assert_equals(null, document.elementFromPoint(viewportX + 100, 10),
                    "X co-ordinates larger than viewport");
                  assert_equals(null, document.elementFromPoint(10, viewportY + 100),
                    "Y co-ordinates larger than viewport");
                  assert_equals(null, document.elementFromPoint(viewportX + 100, viewportY + 100),
                    "X, Y co-ordinates larger than viewport");
    }, "co-ordinates larger than the viewport");

    test(function () {
                  var viewportX = window.frames[1].innerWidth;
                  var viewportY = window.frames[1].innerHeight;
                  var iframeRect = document.getElementById('iframe-2').getBoundingClientRect();
                  assert_equals(null, window.frames[1].document.elementFromPoint(iframeRect.right + viewportX + 100, 10),
                    "X co-ordinates larger than viewport");
                  assert_equals(null, window.frames[1].document.elementFromPoint(10, iframeRect.bottom + viewportY + 100),
                    "Y co-ordinates larger than viewport");
                  assert_equals(null, window.frames[1].document.elementFromPoint(iframeRect.right + viewportX + 100, iframeRect.bottom + viewportY + 100),
                    "X, Y co-ordinates larger than viewport");
    }, "co-ordinates larger than the viewport from in iframe");

    test(function () {
                  assert_equals(document.getElementById('purple'), document.elementFromPoint(10, 10),
                    "Should have returned the element with id `purple`");
                }, "Return first element that is the target for hit testing");

    test(function () {
                  assert_equals(document.getElementById('red'), document.elementFromPoint(10, 120),
                    "Should have returned the element with id `red` as element with `green` has `pointer-events:none`");
                }, "First element to get mouse events with pointer-events css");

    test(function () {
           let svg = document.getElementById('squiggle');
           let svgRect = svg.getBoundingClientRect();
           assert_equals(svg, document.elementFromPoint(svgRect.left + Math.round(svgRect.width/2),
                                                        svgRect.top + Math.round(svgRect.height/2)));
    }, "SVG element at x,y");


    done()
 </script>
